server/requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
pydantic[email]
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator